qPCR data set by Rutledge et al. (2004)

Each data set comprises a six-point, ten-fold dilution series, repeated in five independent runs, for two different amplicons: K1/K2, 102 bp, and K3/K2, 218 bp. Fluorescence readings were exported after background subtraction via baseline averaging of the 5 cycles immediately preceding the cycles in which fluorescence was first detected. Please read the Materials and Methods section of Rutledge et al. (2004) for more details.

Format

A tibble with 10,800 rows and 10 variables:

plate

Plate identifier. Because one plate (run) was used per dilution series, plate values are simply numbered 1 thru 5.

well

Well identifier. Values are always NA (not available). This variable is kept nevertheless to be coherent with other data sets from other similar R data packages.

dye

The type of dye used. In this data set the values are always "SYBR", meaning SYBR Green I master mix (Roche).

target

Target identifier: the amplicon used, K1/K2 or K3/K2.

sample_type

Sample type (all curves are standards, i.e. "std").

replicate

Replicate identifier: 1 thru 4.

copies

Standard copy number.

dilution

Dilution factor. Higher number means greater dilution.

cycle

PCR cycle.

fluor

Raw fluorescence values.

Source

doi:10.1093/nar/gnh177

Examples

rutledge
#> # A tibble: 10,800 × 10
#>    plate well  dye   target sample_type replicate   copies dilution cycle  fluor
#>    <fct> <fct> <chr> <fct>  <fct>           <int>    <int>    <int> <dbl>  <dbl>
#>  1 1     NA    SYBR  K1/K2  std                 1 41700000        1     1 0     
#>  2 1     NA    SYBR  K1/K2  std                 1 41700000        1     2 0     
#>  3 1     NA    SYBR  K1/K2  std                 1 41700000        1     3 0     
#>  4 1     NA    SYBR  K1/K2  std                 1 41700000        1     4 0     
#>  5 1     NA    SYBR  K1/K2  std                 1 41700000        1     5 0.0007
#>  6 1     NA    SYBR  K1/K2  std                 1 41700000        1     6 0.0022
#>  7 1     NA    SYBR  K1/K2  std                 1 41700000        1     7 0.0005
#>  8 1     NA    SYBR  K1/K2  std                 1 41700000        1     8 0.0047
#>  9 1     NA    SYBR  K1/K2  std                 1 41700000        1     9 0.0107
#> 10 1     NA    SYBR  K1/K2  std                 1 41700000        1    10 0.0203
#> # ℹ 10,790 more rows